+++ R/update.R
@@ -6,10 +6,8 @@
 create_df_update <- function() {
     df <- tibble::data_frame("Status" = character(),
                              "Name" = character(),
-                             # Allow for intermediate advisories,
-                             # i.e., "1A", "2", "2A"...
-                             "Adv" = character(),
                              "Date" = as.POSIXct(character(), tz = "UTC"),
+                             "Key" = character(),
                              "Contents" = character())
 
     return(df)
@@ -21,8 +19,8 @@ create_df_update <- function() {
 #'   \item{Status}{Classification of storm, e.g., Tropical Storm, Hurricane,
 #'     etc.}
 #'   \item{Name}{Name of storm}
-#'   \item{Adv}{Advisory Number}
 #'   \item{Date}{Date of advisory issuance}
+#'   \item{Key}{Unique ID of cyclone}
 #'   \item{Contents}{Text content of product}
 #' }
 #' @param link URL to storm's archive page.
@@ -70,18 +68,25 @@ update <- function(link, p = dplyr::progress_estimated(n = 1)) {
 
     status <- scrape_header(contents, ret = "status")
     name <- scrape_header(contents, ret = "name")
-    adv <- scrape_header(contents, ret = "adv")
     date <- scrape_header(contents, ret = "date")
+
+    safely_scrape_header <- purrr::safely(scrape_header)
+    key <- safely_scrape_header(contents, ret = "key")
+    if (is.null(key$error)) {
+        key <- key$result
+    } else {
+        key <- NA
+    }
 
     if (getOption("rrricanes.working_msg"))
         message(sprintf("Working %s %s Update #%s (%s)",
-                        status, name, adv, date))
+                        status, name, date))
 
     df <- df %>%
         tibble::add_row("Status" = status,
                         "Name" = name,
-                        "Adv" = adv,
                         "Date" = date,
+                        "Key" = key,
                         "Contents" = contents)
 
     return(df)