+++ CHANGELOG.md
@@ -4,6 +4,32 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0-4] 2017-06-25
+
+### Added
+  - NA
+
+### Changed
+  - Added variable `Key` to `discus` dataframes. (#80)
+  - Removed variable `Adv` from `posest`. Position estimates do not have advisory numbers. (#81)
+  - Fix `scrape_adv_num` to accomodate possible "INTERMEDIATE" text in Public Advisory headers. (#83)
+  - Remove variable `Adv` from `update`. Updates do not have advisory numbers. (#84)
+  - Added variable `Key` to `get_public` dataframes. (#85)
+  - Added variable `Key` to `get_update` dataframes. (#86)
+  - Removed non-existent wind radii variables in `get_fstadv`. Hrs 48 and 72 hours only have 34 and 50kt wind fields. Hrs 96 and 120 have none. (#89)
+
+### Removed
+  - NA
+
+### Deprecated
+  - NA
+
+### Fixed
+  - NA
+
+### Security
+  - NA
+
 ## [0.2.0-3] 2017-06-22
 
 ### Added